Protect Routes with Authentication
Add this library to read the json web token from the client.
$ npm install express-jwt
Using the express-jwt all routes request are protected with a json web token, unless you specified which ones do not  need to be protected.

If you make any request for example;
http://localhost:3000/places
You will get am UnauthorizedError message like below;

{
    "name": "UnauthorizedError",
    "message": "No authorization token was found",
    "code": "credentials_required",
    "status": 401,
    "inner": {
        "message": "No authorization token was found"
    }
}

But you will to loging or create a user so you need to specified in the unless method that those routes do not need to be protected with a json token.
So if try to login to;
http://localhost:3000/sessions
with the right credentials exmple;
{
	"email": "hello@hello.com",
	"password": "hellohello"

}
Should get the User info and its token without the "UnauthorizedError" protection as below;
{
    "user": {
        "admin": true,
        "_id": "5a627806257fcb26485e77b5",
        "email": "hello@hello.com",
        "name": "Edward",
        "password": "$2a$10$C1xfkdtQmzyKqTQtv9YgMuikSSG/trHVMiK1s6rmoHx9lvNq6ZWzO",
        "__v": 0
    },
    "jwt": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVhNjI3ODA2MjU3ZmNiMjY0ODVlNzdiNSIsImlhdCI6MTUxNjQ3NDQ1NH0.kYNM9GiEMroBYT5PjLjO_tbv_OoFLgjjOJ3OxdMINQc"
}

BUT the route for the places still being protected, you will to use the User token to be able to access to that information.
Using Postman go to http://localhost:3000/places then Headers type 'Authorization' as a key, in the value field you must type the word 'Bearer' before type the right token from the and Authorize User. example;
key: Authorization
value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVhNjI3ODA2MjU3ZmNiMjY0ODVlNzdiNSIsImlhdCI6MTUxNjQ3NDQ1NH0.kYNM9GiEMroBYT5PjLjO_tbv_OoFLgjjOJ3OxdMINQc

If the token is valid you should be able to get the places making a GET request

{
    "docs": [
        {
            "acceptsCreditCard": false,
            "_id": "5a6215b96252780e54a2ed4d",
            "title": "SEO URLS  con Slug",
            "description": "Best company in the world alwa knkn nkn  inkk  ys jnk  ",
            "address": "502west 0001ST New York, NY",
            "openHour": 4,
            "closeHour": 5,
            "slug": "seo-urls-con-slug-3",
            "__v": 0,
            "coverImage": "https://res.cloudinary.com/placesmenteware/image/upload/v1516377527/qrdrahkdkedk9zs05sec.png",
            "avatarImage": "https://res.cloudinary.com/placesmenteware/image/upload/v1516377527/vmfuiuiljj5wub99hcs6.bmp"
        },
        {
            "acceptsCreditCard": false,
            "_id": "5a6215b46252780e54a2ed4b",
            "title": "SEO URLS  con Slug",
            "description": "using multer in express",
            "address": "502west 0001ST New York, NY",
            "openHour": 4,
            "closeHour": 5,
            "slug": "seo-urls-con-slug-1",
            "__v": 0,
            "coverImage": "https://res.cloudinary.com/placesmenteware/image/upload/v1516377522/sl5k363lrlg5ejluyjod.png",
            "avatarImage": "https://res.cloudinary.com/placesmenteware/image/upload/v1516377523/vgycuos1tbmwejvixvni.bmp"
        },
        {
            "acceptsCreditCard": false,
            "_id": "5a6215736252780e54a2ed4a",
            "title": "SEO URLS  con Slug",
            "description": "using multer in express",
            "address": "502west 0001ST New York, NY",
            "openHour": 4,
            "closeHour": 5,
            "slug": "seo-urls-con-slug",
            "__v": 0,
            "coverImage": "https://res.cloudinary.com/placesmenteware/image/upload/v1516377459/bouo57gabqnbaniuwjps.png",
            "avatarImage": "https://res.cloudinary.com/placesmenteware/image/upload/v1516377459/io3qssvhibeubugahldv.bmp"
        }
    ],
    "total": 3,
    "limit": 4,
    "page": 1,
    "pages": 1
}