Virtual to get all Places created by an User
Using a virtual to get all the Places created by an User.
For testins porpuse we uses a GET request along with the User Auth token so in the app.js, we are going to temporary remove 'users' and the method: GET from the unless path of 'express-jwt' only to test the virtual to get all the paces by an User.
Using Postman make GET to http://localhost:3000/users and add the autherization token by an existing authorized User in Header. Should get all the places created by that User.

[
    {
        "acceptsCreditCard": false,
        "_id": "5a63999774c2a100202f6a50",
        "title": "Creado con Auth Json web token",
        "description": "using multer in express",
        "address": "502west 0001ST New York, NY",
        "openHour": 4,
        "closeHour": 5,
        "_user": "5a627806257fcb26485e77b5",
        "slug": "creado-con-auth-json-web-token",
        "__v": 0
    },
    {
        "acceptsCreditCard": false,
        "_id": "5a63a5cc817d582e1cb22939",
        "title": "OTRO Creado con Auth Json web token",
        "description": "using multer in express",
        "address": "502west 0001ST New York, NY",
        "openHour": 4,
        "closeHour": 5,
        "_user": "5a627806257fcb26485e77b5",
        "slug": "otro-creado-con-auth-json-web-token",
        "__v": 0
    },
    {
        "acceptsCreditCard": false,
        "_id": "5a63a5e0817d582e1cb2293a",
        "title": "Tercero OTRO Creado con Auth Json web token",
        "description": "using multer in express",
        "address": "502west 0001ST New York, NY",
        "openHour": 4,
        "closeHour": 5,
        "_user": "5a627806257fcb26485e77b5",
        "slug": "tercero-otro-creado-con-auth-json-web-token",
        "__v": 0
    }
]