Protect Routes for Owner Users to edit and delete
Here we create a middleware to veriry if the user is the owner of the place's document he wants to edit or destroy. if so he can edit his document.
Go to Postman and make a PUT request to the place you want to edit example; http://localhost:3000/places/mi-lugar-para-editar,    and remenber to add the user's token in the Header, if the user created that places, he should be able to edit or destroy the document.