CRUD for Favorites Places
This code is  to Create and Delete favortires Places from an existing Place an Existing User who wants to add the Place to his Favorites Places.
Using Postman send a json POST request in body with the id of the Place you want add to the Favorites Places as below exmple;
{
	"_place": "5a64068c7a32d91724effef8"
}

Also send in the Header add the Autorization token from the Authenticated User who wants to like the Place. exmple;
[{"key":"Authorization","value":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVhNjNlM2RiNTVjYTE1M2QwNGZlMDIxYSIsImlhdCI6MTUxNjU2MDc5Nn0.SGiZy1wgksLNyVpqNo6xs059irDl3-zC3mIjNc2l0h0","description":""}]

Make the POST request and should get the follow info for the User id, the Place id and the FavoritePlace id.

{
    "_id": "5a64e1ccc59aa936583fc20c",
    "_place": "5a64068c7a32d91724effef8",
    "_user": "5a63e3db55ca153d04fe021a",
    "__v": 0
}