Enums for Reactions
Its function is to define a collection of valid predefined data for a field.

An enum is a value type with a set of related named constants often referred to as an enumerator list. The enum keyword is used to declare an enumeration. It is a primitive data type, which is user defined. Enums type can be integer (float, int, byte, double etc.